Don't use state if the display is not available
When we save the state we also save the bounds of the display that hold
the window (or most of the window if it's split on more than 1 display).

When we load, we check if the window will be opened in a display that
has the same bounds as the ones we saved, if not we drop the state and
use the default.