Consider a state without bounds still valid if fullscreen of maximized
The state will have it's coordinates and size info updated only if the
app is not in fullscreen mode or maximized at the time of the update.

This makes sense because coordinates and size are not relevant when in
fullscreen or maximized. The problem is that only the defaultWidth and
defaultHeight definitions will be read if the state in the file is not
considered valid, and to be considered valid, a state must have bounds
(x, y, width and height must be defined).

When leaving the application for the first time (no preexistent
window-store.json), if the application was in fullscreen mode or
maximized when exiting, the newly created window-store.json will not
have coordinates and size info (since the update function that is called
right before the save operation will ignore the coordinates and size, as
said before).

Because of that, when the application is opened for the second time, the
state will be considered invalid and the saved isMaximized or
isFullScreen options will be completely ignored.

This fixes that by considering the state valid if isFullScreen or
isMaximized are true, even if it has no bounds. It also adds some basic
tests to this functionality.